chore: refactor integration tests common functionalities
