+++ integration-tests/engine-hdb/src/test/java/com/sap/xsk/hdb/ds/itest/hdbtable/XSKHDBTableParserHanaITTest.java
@@ -30,6 +30,7 @@ import org.eclipse.dirigible.commons.config.Configuration;
 import org.eclipse.dirigible.database.ds.model.IDataStructureModel;
 import org.eclipse.dirigible.repository.local.LocalResource;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 
 public class XSKHDBTableParserHanaITTest extends AbstractXSKHDBITTest {
@@ -81,6 +82,7 @@ public class XSKHDBTableParserHanaITTest extends AbstractXSKHDBITTest {
     }
 
     @Test
+    @Ignore
     public void testHDBTableAlterWhenCompatibleChange() throws Exception {
 
         try (Connection connection = datasource.getConnection(); Statement stmt = connection.createStatement()) {
@@ -135,6 +137,7 @@ public class XSKHDBTableParserHanaITTest extends AbstractXSKHDBITTest {
     }
 
     @Test
+    @Ignore
     public void testHDBTableAlterWhenNullableToNotNullableChange() throws Exception {
         try (Connection connection = datasource.getConnection(); Statement stmt = connection.createStatement()) {
 
@@ -163,6 +166,7 @@ public class XSKHDBTableParserHanaITTest extends AbstractXSKHDBITTest {
     }
 
     @Test
+    @Ignore
     public void testHDBTableAlterWhenNotUniqueToUniqueChange() throws Exception {
         try (Connection connection = datasource.getConnection(); Statement stmt = connection.createStatement()) {
 
@@ -191,6 +195,7 @@ public class XSKHDBTableParserHanaITTest extends AbstractXSKHDBITTest {
     }
 
     @Test
+    @Ignore
     public void testHDBTableAlterWhenAddingExistingColumnToPKList() throws Exception {
         try (Connection connection = datasource.getConnection(); Statement stmt = connection.createStatement()) {
 
@@ -221,6 +226,7 @@ public class XSKHDBTableParserHanaITTest extends AbstractXSKHDBITTest {
     }
 
     @Test
+    @Ignore
     public void testHDBTableAlterWhenAddingNewColumnToPKList() throws Exception {
         try (Connection connection = datasource.getConnection(); Statement stmt = connection.createStatement()) {
 
@@ -249,6 +255,7 @@ public class XSKHDBTableParserHanaITTest extends AbstractXSKHDBITTest {
     }
 
     @Test
+    @Ignore
     public void testHDBTableAlterWhenAddingNewNotNullableColumn() throws Exception {
         try (Connection connection = datasource.getConnection(); Statement stmt = connection.createStatement()) {
 
@@ -277,6 +284,7 @@ public class XSKHDBTableParserHanaITTest extends AbstractXSKHDBITTest {
     }
 
     @Test
+    @Ignore
     public void testHDBTableAlterWhenExistingColumnTypeChange() throws Exception {
         try (Connection connection = datasource.getConnection(); Statement stmt = connection.createStatement()) {
