+++ integration-tests/pom.xml
@@ -2,26 +2,6 @@
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-  <dependencies>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>${junit.version}</version>
-      <scope>test</scope>
-    </dependency>
-      <dependency>
-          <groupId>org.apache.httpcomponents</groupId>
-          <artifactId>httpmime</artifactId>
-          <version>4.5.13</version>
-          <scope>test</scope>
-      </dependency>
-      <dependency>
-          <groupId>org.apache.httpcomponents</groupId>
-          <artifactId>httpasyncclient</artifactId>
-          <version>4.1.5</version>
-          <scope>test</scope>
-      </dependency>
-  </dependencies>
 
   <parent>
 		<groupId>com.sap.xsk</groupId>
@@ -39,6 +19,7 @@
 		<module>engine-hdb</module>
 		<module>ui</module>
     <module>applications</module>
+    <module>integration-tests-core</module>
   </modules>
 
 	<properties>
