+++ integration-tests/ui/pom.xml
@@ -56,6 +56,12 @@
         </exclusion>
       </exclusions>
     </dependency>
+    <dependency>
+      <groupId>com.sap.xsk</groupId>
+      <artifactId>integration-tests-core</artifactId>
+      <version>0.17.0-SNAPSHOT</version>
+      <scope>test</scope>
+    </dependency>
   </dependencies>
 
   <properties>
